config(test): 모킹 데이터 폴더구조 변경, 장바구니와 상품 목록 네트워크 요청 핸들러 분리
