+++ src/mocks/data/index.ts
@@ -1,0 +1,10 @@
+export { default as ProductsUnfilteredInitial } from "../data/products/basic/products-initial.json";
+export { default as ProductsUnfilteredLast } from "../data/products/basic/products-last.json";
+
+export { default as ProductsUnfilteredSortedDescInitial } from "../data/products/sort/products-sorted-desc-initial.json";
+export { default as ProductsUnfilteredSortedDescLast } from "../data/products/sort/products-sorted-desc-last.json";
+
+export { default as ProductsFilteredBooksInitial } from "../data/products/category/products-filtered-books-initial.json";
+export { default as ProductsFilteredBooksLast } from "../data/products/category/products-filtered-books-last.json";
+export { default as ProductsFilteredBooksSortedAscInitial } from "../data/products/category/products-filtered-books-sorted-asc-initial.json";
+export { default as ProductsFilteredBooksSortedAscLast } from "../data/products/category/products-filtered-books-sorted-asc-last.json";
