+++ src/mocks/data/products/category/products-filtered-books-initial.json
