--- src/mocks/products/index.ts
@@ -1,12 +1,0 @@
-export { default as ProductsUnfilteredInitial } from './products-initial.json';
-export { default as ProductsUnfilteredLast } from './products-last.json';
-export { default as ProductsUnfilteredSortedDescInitial } from './products-sorted-desc-initial.json';
-export { default as ProductsUnfilteredSortedDescLast } from './products-sorted-desc-last.json';
-export { default as ProductsFilteredElectronicsInitial } from './products-filtered-electronics-initial.json';
-export { default as ProductsFilteredElectronicsLast } from './products-filtered-electronics-last.json';
-export { default as ProductsFilteredBooksInitial } from './products-filtered-books-initial.json';
-export { default as ProductsFilteredBooksLast } from './products-filtered-books-last.json';
-export { default as ProductsFilteredBooksSortedAscInitial } from './products-filtered-books-sorted-asc-initial.json';
-export { default as ProductsFilteredBooksSortedAscLast } from './products-filtered-books-sorted-asc-last.json';
-export { default as ProductsFilteredBooksSortedDescInitial } from './products-filtered-books-sorted-desc-initial.json';
-export { default as ProductsFilteredBooksSortedDescLast } from './products-filtered-books-sorted-desc-last.json';
