--- src/mocks/products/products-filtered-books-sorted-desc-initial.json
@@ -1,169 +1,0 @@
-{
-  "content": [
-    {
-      "id": 127,
-      "name": "미술책",
-      "price": 18000,
-      "imageUrl": "https://images.unsplash.com/photo-1543002588-bfa74002ed7e?q=80&w=2730&auto=format&fit=crop&ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8fA%3D%3D",
-      "category": "books"
-    },
-    {
-      "id": 125,
-      "name": "여행책",
-      "price": 15000,
-      "imageUrl": "https://images.unsplash.com/photo-1544716278-ca5e3f4abd8c?q=80&w=3474&auto=format&fit=crop&ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8fA%3D%3D",
-      "category": "books"
-    },
-    {
-      "id": 116,
-      "name": "요리책",
-      "price": 15000,
-      "imageUrl": "https://images.unsplash.com/photo-1589998059171-988d887df646?q=80&w=3552&auto=format&fit=crop&ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8fA%3D%3D",
-      "category": "books"
-    },
-    {
-      "id": 124,
-      "name": "경제서적",
-      "price": 14000,
-      "imageUrl": "https://images.unsplash.com/photo-1589998059171-988d887df646?q=80&w=3552&auto=format&fit=crop&ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8fA%3D%3D",
-      "category": "books"
-    },
-    {
-      "id": 133,
-      "name": "사회과학책",
-      "price": 13000,
-      "imageUrl": "https://plus.unsplash.com/premium_photo-1667251760504-096946b820af?q=80&w=3387&auto=format&fit=crop&ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8fA%3D%3D",
-      "category": "books"
-    },
-    {
-      "id": 117,
-      "name": "그래픽노블",
-      "price": 13000,
-      "imageUrl": "https://images.unsplash.com/photo-1544716278-ca5e3f4abd8c?q=80&w=3474&auto=format&fit=crop&ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8fA%3D%3D",
-      "category": "books"
-    },
-    {
-      "id": 123,
-      "name": "과학책",
-      "price": 13000,
-      "imageUrl": "https://plus.unsplash.com/premium_photo-1667251760504-096946b820af?q=80&w=3387&auto=format&fit=crop&ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8fA%3D%3D",
-      "category": "books"
-    },
-    {
-      "id": 131,
-      "name": "역사논문",
-      "price": 12000,
-      "imageUrl": "https://images.unsplash.com/photo-1589998059171-988d887df646?q=80&w=3552&auto=format&fit=crop&ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8fA%3D%3D",
-      "category": "books"
-    },
-    {
-      "id": 126,
-      "name": "음악책",
-      "price": 12000,
-      "imageUrl": "https://plus.unsplash.com/premium_photo-1681825268400-c561bd47d586?q=80&w=3544&auto=format&fit=crop&ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8fA%3D%3D",
-      "category": "books"
-    },
-    {
-      "id": 115,
-      "name": "자기계발서",
-      "price": 12000,
-      "imageUrl": "https://plus.unsplash.com/premium_photo-1667251760504-096946b820af?q=80&w=3387&auto=format&fit=crop&ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8fA%3D%3D",
-      "category": "books"
-    },
-    {
-      "id": 118,
-      "name": "역사책",
-      "price": 12000,
-      "imageUrl": "https://plus.unsplash.com/premium_photo-1681825268400-c561bd47d586?q=80&w=3544&auto=format&fit=crop&ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8fA%3D%3D",
-      "category": "books"
-    },
-    {
-      "id": 132,
-      "name": "철학논문",
-      "price": 11000,
-      "imageUrl": "https://images.unsplash.com/photo-1544716278-ca5e3f4abd8c?q=80&w=3474&auto=format&fit=crop&ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8fA%3D%3D",
-      "category": "books"
-    },
-    {
-      "id": 122,
-      "name": "철학책",
-      "price": 11000,
-      "imageUrl": "https://images.unsplash.com/photo-1544716278-ca5e3f4abd8c?q=80&w=3474&auto=format&fit=crop&ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8fA%3D%3D",
-      "category": "books"
-    },
-    {
-      "id": 121,
-      "name": "에세이",
-      "price": 10000,
-      "imageUrl": "https://images.unsplash.com/photo-1589998059171-988d887df646?q=80&w=3552&auto=format&fit=crop&ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8fA%3D%3D",
-      "category": "books"
-    },
-    {
-      "id": 111,
-      "name": "자서전",
-      "price": 10000,
-      "imageUrl": "https://images.unsplash.com/photo-1589998059171-988d887df646?q=80&w=3552&auto=format&fit=crop&ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8fA%3D%3D",
-      "category": "books"
-    },
-    {
-      "id": 119,
-      "name": "동화책",
-      "price": 10000,
-      "imageUrl": "https://images.unsplash.com/photo-1543002588-bfa74002ed7e?q=80&w=2730&auto=format&fit=crop&ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8fA%3D%3D",
-      "category": "books"
-    },
-    {
-      "id": 128,
-      "name": "문학소설",
-      "price": 9000,
-      "imageUrl": "https://plus.unsplash.com/premium_photo-1681825268400-c561bd47d586?q=80&w=3544&auto=format&fit=crop&ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8fA%3D%3D",
-      "category": "books"
-    },
-    {
-      "id": 114,
-      "name": "스릴러소설",
-      "price": 9000,
-      "imageUrl": "https://images.unsplash.com/photo-1543002588-bfa74002ed7e?q=80&w=2730&auto=format&fit=crop&ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8fA%3D%3D",
-      "category": "books"
-    },
-    {
-      "id": 129,
-      "name": "고전소설",
-      "price": 8000,
-      "imageUrl": "https://images.unsplash.com/photo-1543002588-bfa74002ed7e?q=80&w=2730&auto=format&fit=crop&ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8fA%3D%3D",
-      "category": "books"
-    },
-    {
-      "id": 120,
-      "name": "시집",
-      "price": 8000,
-      "imageUrl": "https://plus.unsplash.com/premium_photo-1667251760504-096946b820af?q=80&w=3387&auto=format&fit=crop&ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8fA%3D%3D",
-      "category": "books"
-    }
-  ],
-  "pageable": {
-    "sort": {
-      "sorted": true,
-      "unsorted": false,
-      "empty": false
-    },
-    "pageNumber": 0,
-    "pageSize": 20,
-    "offset": 0,
-    "paged": true,
-    "unpaged": false
-  },
-  "last": false,
-  "totalPages": 2,
-  "totalElements": 24,
-  "sort": {
-    "sorted": true,
-    "unsorted": false,
-    "empty": false
-  },
-  "first": true,
-  "number": 0,
-  "numberOfElements": 20,
-  "size": 20,
-  "empty": false
-}
